ReportServiceStatus fixes

+++ src/main/java/com/ewp/crm/service/interfaces/ReportService.java
@@ -7,6 +7,6 @@ import java.util.Set;
 
 public interface ReportService {
     int countChangedStatusClients(ZonedDateTime firstReportDate, ZonedDateTime lastReportDate, Status from, Status to, Set<Status> exclude);
-    int countNewClients(ZonedDateTime firstReportDate, ZonedDateTime lastReportDate);
+    long countNewClients(ZonedDateTime firstReportDate, ZonedDateTime lastReportDate);
     long countFirstPaymentClients(Status inProgressStatus, ZonedDateTime firstReportDate, ZonedDateTime lastReportDate);
 }
